new categories added

+++ src/com/ut/trivia/starter/AppStarter.java
@@ -27,21 +27,27 @@ public class AppStarter {
             Scanner scanner = new Scanner(System.in);
             int userChoice = GeneralConstants.MINUS_ONE_INT;
         	do {
+        		HandleIpOp.printGeneralMode();
         		userChoice = HandleIpOp.getCategoryChoice(scanner);
         		if(userChoice == AdminConstants.ADMIN_CODE) { // if admin mode is activated
             		HandleIpOp.printAdminWelcome();
             		int adminOperationChoice;
-            		do{ // while admin wants to perform certain tasks
+            		int adminCategory;
+            		do {
             			HandleIpOp.printAdminMode();
-            			int adminCategory = HandleIpOp.getAdminCategoryChoice(scanner);
-                		adminOperationChoice = HandleIpOp.getAdminOperationChoice(scanner);
-                		switch (adminOperationChoice){ // switch for admin operations
-                		case 1: AdminFunctionalitiesImpl.addQuestion(adminCategory,scanner);break;
-                		case 2: AdminFunctionalitiesImpl.editQuestion(adminCategory,scanner); break;
-                		case 3: AdminFunctionalitiesImpl.deleteQuestion(adminCategory,scanner); break;
-                		case 4: break;
+                		adminCategory = HandleIpOp.getAdminCategoryChoice(scanner);
+                		if(adminCategory != AppData.getCategories().size()+1) {
+                			do{ // while admin wants to perform certain tasks
+                        		adminOperationChoice = HandleIpOp.getAdminOperationChoice(scanner);
+                        		switch (adminOperationChoice){ // switch for admin operations
+                        		case 1: AdminFunctionalitiesImpl.addQuestion(adminCategory,scanner);break;
+                        		case 2: AdminFunctionalitiesImpl.editQuestion(adminCategory,scanner); break;
+                        		case 3: AdminFunctionalitiesImpl.deleteQuestion(adminCategory,scanner); break;
+                        		case 4: break;
+                        		}
+                    		}while(adminOperationChoice != GeneralConstants.FOUR_INT);
                 		}
-            		}while(adminOperationChoice != GeneralConstants.FOUR_INT);
+            		}while(adminCategory != AppData.getCategories().size()+1);
             	}else if(userChoice == AppData.getCategories().size()+1){ // if user wants to exit
             		break;
             	}else { // if user mode is activated
